pyproject.toml
[tool.poetry]
name = "lark-experiments"
version = "1.0.0-alpha.0"
description = ""
authors = ["Drew O'Kane <rhoslug@gmail.com>"]
repository="https://github.com/drewokane/wooble-lang"

[tool.poetry.dependencies]
python = "*"
lark-parser = "^0.6.4"
ipython = "^6.5"
toolz = "^0.9.0"
pytest-cov = "^2.6"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
